+++ src/main/java/baseball/game/Computer.java
@@ -1,4 +1,4 @@
-package baseball.computer;
+package baseball.game;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
